ingestion/requirements.txt
fastapi
uvicorn[standard]
kafka-python
pydantic
numpy
